Upload Click File Explorer
The plus sign now opens up file explorer when clicked. However, we still need to fix the main page. The click is still registered if you click on the left and right side outisde of the upload box which is unacceptable. Also, we need to shrink the plus sign and verify that the screen is responsive and consistent on different screens/resolutions.