GUI Changes
Made it so that the middle container no longer has a gap at the top of it so that you can see the baby blue background for all the footers. Changed the background color of the footer and copyright containers and the text color of the content within that container. Left comments on the styleMain.css file where you can experiment with the background colors.